project2/requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
spacy
pydantic
python-multipart
Pillow
pytest 